requirements.txt
fastapi>=0.116
uvicorn[standard]>=0.29
pydantic>=2.7
sqlmodel>=0.0.16
pandas>=2.2
numpy>=2.3
statsmodels>=0.14
polygon-api-client>=1.12
yfinance>=0.2
websockets>=12.0
plotly>=5.22
python-multipart>=0.0.9
